requirements.txt
Django==3.2.1
django-materializecss-form==1.1.17
django-multiselectfield==0.1.12
django-dotenv==1.4.2
